refactor(domain): move EvaluationOutcome + EvaluationResult into voom-domain with serde derives (#378)
