+++ crates/voom-domain/src/lib.rs
@@ -5,6 +5,7 @@ pub mod capabilities;
 pub mod capability_map;
 pub mod capability_resolution;
 pub mod compiled;
+pub mod evaluation;
 pub mod errors;
 pub mod estimate;
 pub mod events;
